[refactor] Project와 ProjectUser 패키지 분리
