+++ module-service/src/main/java/com/welcommu/moduleservice/project/dto/ProjectSummaryResponse.java
@@ -1,8 +1,8 @@
 package com.welcommu.moduleservice.project.dto;
 
 import com.welcommu.moduledomain.project.Project;
-import com.welcommu.moduledomain.project.ProjectUser;
-import com.welcommu.moduledomain.project.ProjectUserManageRole;
+import com.welcommu.moduledomain.projectUser.ProjectUser;
+import com.welcommu.moduledomain.projectUser.ProjectUserManageRole;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
