+++ module-service/src/main/java/com/welcommu/moduleservice/project/dto/ProjectUserRoleRequest.java
@@ -1,7 +1,8 @@
 package com.welcommu.moduleservice.project.dto;
 
 import com.welcommu.moduledomain.project.Project;
-import com.welcommu.moduledomain.project.ProjectUser;
+import com.welcommu.moduledomain.projectUser.ProjectUser;
+import com.welcommu.moduledomain.projectUser.ProjectUserManageRole;
 import com.welcommu.moduledomain.user.User;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -12,7 +13,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public class ProjectUserRoleRequest {
     private Long userId;
-    private com.welcommu.moduledomain.project.ProjectUserManageRole role;
+    private ProjectUserManageRole role;
 
     public ProjectUser toEntity(Project project, User user) {
         return ProjectUser.builder()
